Split weekly report file into a directory
