+++ src/sentry/tasks/reports/backends/dummy.py
@@ -1,0 +1,12 @@
+from functools import partial
+
+from .base import ReportBackend
+
+
+class DummyReportBackend(ReportBackend):
+    def prepare(self, timestamp, duration, organization):
+        pass
+
+    def fetch(self, timestamp, duration, organization, projects):
+        assert all(project.organization_id == organization.id for project in projects)
+        return map(partial(self.build, timestamp, duration), projects)
