+++ src/sentry/tasks/reports/types/__init__.py
@@ -1,0 +1,11 @@
+from .distribution import DistributionType
+from .duration import Duration
+from .key import Key
+from .skipped import Skipped
+
+__all__ = (
+    "DistributionType",
+    "Duration",
+    "Key",
+    "Skipped",
+)
