+++ src/sentry/tasks/reports/types/distribution.py
@@ -1,0 +1,6 @@
+from typing import NamedTuple
+
+
+class DistributionType(NamedTuple):
+    label: str
+    color: str
