+++ src/sentry/tasks/reports/types/key.py
@@ -1,0 +1,8 @@
+from typing import Mapping, NamedTuple, Optional
+
+
+class Key(NamedTuple):
+    label: str
+    url: Optional[str]
+    color: str
+    data: Mapping[str, int]
