+++ src/sentry/tasks/reports/types/skipped.py
@@ -1,0 +1,4 @@
+class Skipped:
+    NotSubscribed = object()
+    NoProjects = object()
+    NoReports = object()
